requirements.txt
blessed==1.20.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
gpxpy==1.5.0
inquirer==3.1.3
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
readchar==4.0.5
scipy==1.10.1
six==1.16.0
wcwidth==0.2.6
